Move no owner toggle to the list title
