+++ src/js/views/patients/schedule/layout.hbs
@@ -5,7 +5,7 @@
   </div>
   <div class="list-page__filters flex">
     <div class="flex u-margin--r-24" data-select-all-region></div>
-    <div class="w-100" data-filters-region></div>
+    <div class="flex flex-align-center w-100" data-filters-region></div>
     <div class="flex u-text--nowrap" data-date-filter-region></div>
   </div>
   <table class="w-100" data-table-region></table>
