+++ src/js/views/patients/worklist/layout.hbs
@@ -6,7 +6,7 @@
   </div>
   <div class="list-page__filters flex">
     <div class="flex u-margin--r-24" data-select-all-region></div>
-    <div class="worklist-list__filter-region" data-filters-region></div>
+    <div class="flex flex-align-center w-100" data-filters-region></div>
     <div class="flex">
       <span class="u-margin--r-16 u-text--nowrap" data-date-filter-region></span>
       <span class="worklist-list__filter-sort" data-sort-region></span>
